refactor: extract SettingsContent components to comply with 50-line rule
