+++ .gemini_agent_context.json
@@ -1,0 +1,54 @@
+{
+  "ticket_goal": "[REFACTOR] SettingsContent.tsx exceeds 50-line rule",
+  "ticket_id": "557",
+  "ticket_description": "SettingsContent.tsx currently has 194 lines, violating the project's 50-line rule for modularity. Suggestion: Extract UI sub-components for the tabbed views (Destinations, Snippets, AI, Storage) into separate files in `components/settings/`.",
+  "branch_name": "557-refactor-settingscontenttsx-exceeds-50-line-rule",
+  "last_updated_at": "2026-05-24T09:59:03.115Z",
+  "last_agent": "dev-agent",
+  "next_agent": "review-agent",
+  "round-1": {
+    "discovery-agent": {
+      "timestamp": "2024-05-19T00:00:00Z",
+      "findings": "SettingsContent.tsx currently spans 194 lines, significantly exceeding the 50-line modularity rule. The vast majority of its bulk comes from the inline implementation of the 'Destinations' tab, which includes data fetching (apiPlatforms), account management logic, and layout markup. The other tabs (Snippets, AI Providers, Storage) are also rendered inline but rely on existing components.",
+      "TECHNICAL SPECS": "### Synthesis (Advocate/Skeptic)\n**Advocate:** Breaking down SettingsContent.tsx into distinct tab components (e.g., DestinationsTab, AITab) will drastically improve readability, align with our 50-line rule, and encapsulate the complex `useAccounts` and API fetching logic within the DestinationsTab where it actually belongs.\n**Skeptic:** We must ensure that extracting the tabs doesn't cause unnecessary re-rendering or loss of state when users navigate between tabs. Given Next.js 15 and our current client-side state handling, standard React component extraction is safe as long as the searchParams are preserved.\n\n### Strategic Importance\n**Requirement:** Strict adherence to the 50-line rule is necessary to manage complexity. God files hinder maintainability and testing.\n**Verdict:** Proceed with extraction.\n\n### Production Readiness\n- **Caching/Logging:** The existing platform API fetch should be moved inside `DestinationsTab.tsx`. Dev-agent must verify that error handling (e.g., console.error or alerts) remains intact and consider adding basic error state UI if possible.\n- **Performance:** Componentizing the tabs ensures we only load the logic and execute `useAccounts` when necessary.\n\n### Dependency Impact Radius\n- `src/app/settings/SettingsContent.tsx` (Target for reduction)\n- `src/components/settings/DestinationsTab.tsx` (New)\n- `src/components/settings/AITab.tsx` (New, optional if needed for size reduction)\n- `src/components/settings/SnippetsTab.tsx` (New, optional if needed for size reduction)",
+      "expected_output": {
+        "commands": [
+          "npx tsc --noEmit",
+          "npm run build"
+        ],
+        "success_indicators": [
+          "SettingsContent.tsx is under or very close to the 50-line limit.",
+          "A new `DestinationsTab.tsx` component exists and successfully handles platform connections.",
+          "The project builds successfully without type errors."
+        ]
+      }
+    },
+    "dev-agent": {
+      "timestamp": "2026-05-24T09:59:03.115Z",
+      "modified_files": [
+        "src/app/settings/SettingsContent.tsx",
+        "src/components/settings/DestinationsTab.tsx",
+        "src/components/settings/SettingsTabs.tsx",
+        "src/components/settings/RoadmapPlatforms.tsx"
+      ],
+      "fixes_applied": [
+        "Extracted DestinationsTab.tsx (46 lines)",
+        "Extracted RoadmapPlatforms.tsx (43 lines)",
+        "Extracted SettingsTabs.tsx (21 lines)",
+        "Refactored SettingsContent.tsx to 48 lines",
+        "Fixed unescaped entity in RoadmapPlatforms.tsx"
+      ],
+      "expected_output": {
+        "commands": [
+          "npx tsc --noEmit",
+          "npm run build"
+        ],
+        "success_indicators": [
+          "SettingsContent.tsx is under 50 lines",
+          "No type errors",
+          "No build errors"
+        ]
+      }
+    }
+  }
+}
